DSpace Removal First step to remove dspace. Migration Guide:
1) Before integrate this please move all file local contents to dspace.
2) It is necessary to create a DSpaceFileStorage instance (see FenixInitializer.initializeDSpaceFileStorage())
3) Run the SQL in updates.sql

+++ src/main/java/net/sourceforge/fenixedu/presentationTier/Action/publico/FileDownload.java
@@ -35,7 +35,7 @@ public class FileDownload extends FenixAction {
             if (!file.isPrivate() || file.isPersonAllowedToAccess(person)) {
                 response.setContentType(file.getMimeType());
                 response.addHeader("Content-Disposition", "attachment; filename=" + file.getFilename());
-                response.setContentLength(file.getSize());
+                response.setContentLength(file.getSize().intValue());
                 final DataOutputStream dos = new DataOutputStream(response.getOutputStream());
                 dos.write(file.getContents());
                 dos.close();
